Moved a few functions to util.rs
