+++ .gitignore
@@ -1,2 +1,3 @@
 target/
 juicebox.code-workspace
+documents-of-interest/
