+++ src/main.rs
@@ -13,7 +13,7 @@ use simulation::sim_physics_engine;
 fn main() {
     let mut juicebox: App = App::new();
 	juicebox.add_plugins((
-		DefaultPlugins.set(juice_renderer::create_window_plugin()),
+		DefaultPlugins.set(util::create_window_plugin()),
 		juice_renderer::JuiceRenderer,
 		sim_state_manager::SimStateManager,
 		sim_physics_engine::SimPhysicsEngine,
